claude-seo-unified-complete/scripts/run_skill_workflow.py
#!/usr/bin/env python3
"""
Headless workflow executor for claude-seo-unified
Enables deterministic execution in CI/CD, cron, or API mode
"""

import sys
import json
import argparse
from pathlib import Path
from typing import Dict, Any

def load_config(config_path: str = "config/config.yaml") -> Dict[str, Any]:
    """Load YAML config"""
    import yaml
    with open(config_path) as f:
        return yaml.safe_load(f)

def run_audit(url: str, output_format: str = "json") -> Dict[str, Any]:
    """Execute SEO audit workflow"""
    # This would call the actual audit orchestrator
    # Placeholder for demonstration
    return {
        "url": url,
        "timestamp": "2026-04-28T14:30:22Z",
        "health_score": 73,
        "scores": {
            "technical": 18,
            "content": 19,
            "onpage": 16,
            "schema": 8,
            "performance": 7,
            "ai_readiness": 4,
            "images": 3
        },
        "issues": [
            {"severity": "high", "category": "technical", "issue": "Missing robots.txt"},
            {"severity": "medium", "category": "schema", "issue": "No Organization schema"},
            {"severity": "low", "category": "performance", "issue": "LCP 2.8s"}
        ],
        "recommendations": [
            "Add robots.txt file",
            "Implement Organization schema",
            "Optimize largest contentful paint"
        ]
    }

def run_drift_compare(baseline_path: str, current_url: str) -> Dict[str, Any]:
    """Compare current state vs baseline"""
    return {
        "baseline_score": 81,
        "current_score": 73,
        "delta": -8,
        "regressions": [
            {"metric": "LCP", "baseline": 2.1, "current": 2.8, "change": "+33%"},
            {"metric": "Schema coverage", "baseline": 10, "current": 8, "change": "-20%"}
        ],
        "improvements": [],
        "timestamp": "2026-04-28T14:30:22Z"
    }

def main():
    parser = argparse.ArgumentParser(description="Run SEO workflow headless")
    parser.add_argument("workflow", choices=["audit", "drift", "competitor", "monitor"],
                       help="Workflow to execute")
    parser.add_argument("--url", required=True, help="Target URL")
    parser.add_argument("--baseline", help="Baseline file for drift comparison")
    parser.add_argument("--format", choices=["json", "yaml", "pdf"], default="json",
                       help="Output format")
    parser.add_argument("--output", help="Output file path")
    parser.add_argument("--config", default="config/config.yaml", help="Config file")
    
    args = parser.parse_args()
    
    # Load config
    config = load_config(args.config)
    
    # Execute workflow
    if args.workflow == "audit":
        result = run_audit(args.url, args.format)
    elif args.workflow == "drift":
        if not args.baseline:
            print("Error: --baseline required for drift workflow", file=sys.stderr)
            return 1
        result = run_drift_compare(args.baseline, args.url)
    else:
        print(f"Workflow '{args.workflow}' not yet implemented", file=sys.stderr)
        return 1
    
    # Output result
    if args.output:
        output_path = Path(args.output)
        output_path.parent.mkdir(parents=True, exist_ok=True)
        
        if args.format == "json":
            with open(output_path, "w") as f:
                json.dump(result, f, indent=2)
        elif args.format == "yaml":
            import yaml
            with open(output_path, "w") as f:
                yaml.dump(result, f)
        else:  # pdf
            print("PDF generation requires full audit workflow", file=sys.stderr)
            return 1
            
        print(f"✅ Result saved to {output_path}")
    else:
        print(json.dumps(result, indent=2))
    
    return 0

if __name__ == "__main__":
    sys.exit(main())
