claude-seo-unified-complete/requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0

# Web scraping & rendering
playwright>=1.40.0
selenium>=4.15.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# PDF/Report generation
weasyprint>=60.0
matplotlib>=3.8.0
pillow>=10.0.0

# Schema validation
jsonschema>=4.19.0

# Optional (MCP integrations)
anthropic>=0.7.0  # For Claude API
openai>=1.3.0     # For GPT integrations

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
